// PersonalWebsite-main/src/common/types.tsx
export type navItemProps = { link: string; value: string };

export type experienceProps = {
  company: string;
  url: string;
  role: string;
  timeline: string;
  bullets: string[];
  location: string;
};

export type projectProps = {
  name: string;
  skills: string[];
  timeline: string;
  text: string;
  url: string;
};

export type aboutProps = {
  topic: string;
  role: string;
  bullets: string[];
};